requirements.txt
packages/chatgpt_wrapper-1.0-py3-none-any.whl
discord.py
python-dotenv